Add files via upload
A simple digit recognition app built with Pygame. Users can draw digits on the screen, and the app uses the trained CNN model (bestmodel.h5) to predict and display the digit above a red bounding box. Input is processed to match MNIST format. Press "n" to clear the screen and draw again.